+++ src/media.rs
@@ -1,4 +1,5 @@
 pub(crate) mod audio_data;
+pub(crate) mod custom_audio_source;
 pub(crate) mod native_vad;
 pub(crate) mod video_frame;
 pub(crate) mod virtual_camera_device;
@@ -6,6 +7,7 @@ pub(crate) mod virtual_microphone_device;
 pub(crate) mod virtual_speaker_device;
 
 pub(crate) use audio_data::PyAudioData;
+pub(crate) use custom_audio_source::PyCustomAudioSource;
 pub(crate) use native_vad::PyNativeVad;
 pub(crate) use video_frame::PyVideoFrame;
 pub(crate) use virtual_camera_device::PyVirtualCameraDevice;